fix: SectionItems
